+++ modules/Students/includes/Enrollment.inc.php
@@ -64,7 +64,7 @@ $calendars_RET = DBGet( "SELECT CALENDAR_ID,DEFAULT_CALENDAR,TITLE
 	FROM attendance_calendars
 	WHERE SYEAR='" . UserSyear() . "'
 	AND SCHOOL_ID='" . UserSchool() . "'
-	ORDER BY DEFAULT_CALENDAR ASC" );
+	ORDER BY DEFAULT_CALENDAR IS NULL,DEFAULT_CALENDAR ASC,TITLE" );
 
 $calendar_options = [];